IA-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
python-jose[cryptography]
pydantic
python-docx
pytest
httpx